replace rust_decimal with bigdecimal for arbitrary precision
This commit updates the version of the uniswap-sdk-core-rust library to 0.5.0 and includes bigdecimal as a dependency, replacing the previous rust_decimal. The major change involves the update in the Fraction and CurrencyAmount where BigDecimal is now used for calculations instead of Decimal from rust_decimal. The relevant changes in the code are also refactored to fit with the bigdecimal usage. The Cargo.lock file has also been updated to reflect these changes.