Upgrade alloy-primitives dependency to 0.6.0 and refactor fraction implementation
This commit upgrades the alloy-primitives dependency version from 0.5.4 to 0.6.0 in the Cargo.toml and Cargo.lock files. It also includes a significant refactor of the fraction implementation, by moving from FractionTrait to FractionBase and improving the numerator and denominator methods in the FractionTrait. The changes aim to make the fraction implementation more intuitive and efficient.