feat(macro): Update token macro and fix address formatting
The token macro in `src/entities/token.rs` has been updated to create tokens with more levels of specificity based on the passed arguments. This includes creating tokens with only chain_id, address, and decimals, as well as tokens with additional symbol and name parameters. The address formatting in `src/entities/weth9.rs` is modified to remove leading zeroes for clearer representation. Also, the package is bumped to version `0.17.1` in `Cargo.toml` due to these changes.